refactor: Move methods to appropriate files
- Utility methods to utils.py
- Methods accessible through via API to api.py
