+++ frontend/src/store.ts
@@ -312,7 +312,7 @@ const useStore = defineStore("store", {
 				private: false,
 				folder: "Home/Builder Uploads",
 				optimize: true,
-				upload_endpoint: "/api/method/builder.builder.doctype.builder_page.builder_page.upload_builder_asset",
+				upload_endpoint: "/api/method/builder.api.upload_builder_asset",
 			});
 			await new Promise((resolve) => {
 				toast.promise(upload, {
